Moved base command class into commands module
